[PATCH] ipmi_kcs_intf.c compile warning
From: Zwane Mwaikambo <zwane@linuxpower.ca>

drivers/char/ipmi/ipmi_kcs_intf.c: In function `acpi_find_bmc':
drivers/char/ipmi/ipmi_kcs_intf.c:1088: warning: long unsigned int format, different type arg (arg 2)
drivers/char/ipmi/ipmi_kcs_intf.c:1088: warning: long unsigned int format, different type arg (arg 2)

+++ drivers/char/ipmi/ipmi_kcs_intf.c
@@ -1085,7 +1085,7 @@ static int acpi_find_bmc(unsigned long *physaddr, int *port)
 		*port = spmi->addr.address;
 		printk("ipmi_kcs_intf: Found ACPI-specified state machine"
 		       " at I/O address 0x%lx\n",
-		       (int) spmi->addr.address);
+		       (unsigned long) spmi->addr.address);
 	} else
 		goto not_found; /* Not an address type we recognise. */
